+++ airbyte-cdk/python/unit_tests/sources/file_based/scenarios/parquet_scenarios.py
@@ -170,9 +170,9 @@ single_parquet_scenario = (
     .set_file_type("parquet")
     .set_expected_records(
         [
-            {"data": {"col1": "val11", "col2": "val12", "_ab_source_file_last_modified": "2023-06-05T03:54:07Z",
+            {"data": {"col1": "val11", "col2": "val12", "_ab_source_file_last_modified": "2023-06-05T03:54:07.000000Z",
                       "_ab_source_file_url": "a.parquet"}, "stream": "stream1"},
-            {"data": {"col1": "val21", "col2": "val22", "_ab_source_file_last_modified": "2023-06-05T03:54:07Z",
+            {"data": {"col1": "val21", "col2": "val22", "_ab_source_file_last_modified": "2023-06-05T03:54:07.000000Z",
                       "_ab_source_file_url": "a.parquet"}, "stream": "stream1"},
         ]
     )
@@ -185,10 +185,10 @@ single_parquet_scenario = (
                         "type": "object",
                         "properties": {
                             "col1": {
-                                "type": "string"
+                                "type": ["null", "string"]
                             },
                             "col2": {
-                                "type": "string"
+                                "type": ["null", "string"]
                             },
                             "_ab_source_file_last_modified": {
                                 "type": "string"
@@ -233,13 +233,13 @@ multi_parquet_scenario = (
                         "type": "object",
                         "properties": {
                             "col1": {
-                                "type": "string"
+                                "type": ["null", "string"]
                             },
                             "col2": {
-                                "type": "string"
+                                "type": ["null", "string"]
                             },
                             "col3": {
-                                "type": "string"
+                                "type": ["null", "string"]
                             },
                             "_ab_source_file_last_modified": {
                                 "type": "string"
@@ -258,13 +258,13 @@ multi_parquet_scenario = (
     )
     .set_expected_records(
         [
-            {"data": {"col1": "val11a", "col2": "val12a", "_ab_source_file_last_modified": "2023-06-05T03:54:07Z",
+            {"data": {"col1": "val11a", "col2": "val12a", "_ab_source_file_last_modified": "2023-06-05T03:54:07.000000Z",
                       "_ab_source_file_url": "a.parquet"}, "stream": "stream1"},
-            {"data": {"col1": "val21a", "col2": "val22a", "_ab_source_file_last_modified": "2023-06-05T03:54:07Z",
+            {"data": {"col1": "val21a", "col2": "val22a", "_ab_source_file_last_modified": "2023-06-05T03:54:07.000000Z",
                       "_ab_source_file_url": "a.parquet"}, "stream": "stream1"},
-            {"data": {"col1": "val11b", "col2": "val12b", "col3": "val13b", "_ab_source_file_last_modified": "2023-06-05T03:54:07Z",
+            {"data": {"col1": "val11b", "col2": "val12b", "col3": "val13b", "_ab_source_file_last_modified": "2023-06-05T03:54:07.000000Z",
                       "_ab_source_file_url": "b.parquet"}, "stream": "stream1"},
-            {"data": {"col1": "val21b", "col2": "val22b", "col3": "val23b", "_ab_source_file_last_modified": "2023-06-05T03:54:07Z",
+            {"data": {"col1": "val21b", "col2": "val22b", "col3": "val23b", "_ab_source_file_last_modified": "2023-06-05T03:54:07.000000Z",
                       "_ab_source_file_url": "b.parquet"}, "stream": "stream1"},
         ]
     )
@@ -296,80 +296,80 @@ parquet_various_types_scenario = (
                         "type": "object",
                         "properties": {
                             "col_bool": {
-                                "type": "boolean"
+                                "type": ["null", "boolean"],
                             },
                             "col_int8": {
-                                "type": "integer"
+                                "type": ["null", "integer"],
                             },
                             "col_int16": {
-                                "type": "integer"
+                                "type": ["null", "integer"],
                             },
                             "col_int32": {
-                                "type": "integer"
+                                "type": ["null", "integer"],
                             },
                             "col_uint8": {
-                                "type": "integer"
+                                "type": ["null", "integer"],
                             },
                             "col_uint16": {
-                                "type": "integer"
+                                "type": ["null", "integer"],
                             },
                             "col_uint32": {
-                                "type": "integer"
+                                "type": ["null", "integer"],
                             },
                             "col_uint64": {
-                                "type": "integer"
+                                "type": ["null", "integer"],
                             },
                             "col_float32": {
-                                "type": "number"
+                                "type": ["null", "number"],
                             },
                             "col_float64": {
-                                "type": "number"
+                                "type": ["null", "number"],
                             },
                             "col_string": {
-                                "type": "string"
+                                "type": ["null", "string"],
                             },
                             "col_date32": {
-                                "type": "string",
+                                "type": ["null", "string"],
                                 "format": "date"
                             },
                             "col_date64": {
-                                "type": "string",
+                                "type": ["null", "string"],
                                 "format": "date"
                             },
                             "col_timestamp_without_tz": {
-                                "type": "string",
+                                "type": ["null", "string"],
                                 "format": "date-time"
                             },
                             "col_timestamp_with_tz": {
-                                "type": "string",
+                                "type": ["null", "string"],
                                 "format": "date-time"
                             },
                             "col_time32s": {
-                                "type": "string",
+                                "type": ["null", "string"],
                             },
                             "col_time32ms": {
-                                "type": "string",
+                                "type": ["null", "string"],
                             },
                             "col_time64us": {
-                                "type": "string",
+                                "type": ["null", "string"],
                             },
                             "col_struct": {
-                                "type": "object",
+                                "type": ["null", "object"],
                             },
                             "col_list": {
-                                "type": "array",
+                                "type": ["null", "array"],
                             },
                             "col_duration": {
-                                "type": "integer",
+                                "type": ["null", "integer"],
                             },
                             "col_binary": {
-                                "type": "string",
+                                "type": ["null", "string"],
                             },
                             "_ab_source_file_last_modified": {
-                                "type": "string"
+                                "type": "string",
                             },
                             "_ab_source_file_url": {
-                                "type": "string"
+                                "type": "string",
                             },
                         },
                     },
@@ -404,7 +404,7 @@ parquet_various_types_scenario = (
                       "col_list": [1, 2, 3, 4],
                       "col_duration": 12345,
                       "col_binary": "binary string. Hello world!",
-                      "_ab_source_file_last_modified": "2023-06-05T03:54:07Z",
+                      "_ab_source_file_last_modified": "2023-06-05T03:54:07.000000Z",
                       "_ab_source_file_url": "a.parquet"}, "stream": "stream1"
              },
         ]
@@ -430,7 +430,7 @@ parquet_file_with_decimal_no_config_scenario = (
     .set_file_type("parquet")
     .set_expected_records(
         [
-            {"data": {"col1": "13.00", "_ab_source_file_last_modified": "2023-06-05T03:54:07Z",
+            {"data": {"col1": "13.00", "_ab_source_file_last_modified": "2023-06-05T03:54:07.000000Z",
                       "_ab_source_file_url": "a.parquet"}, "stream": "stream1"},
         ]
     )
@@ -443,7 +443,7 @@ parquet_file_with_decimal_no_config_scenario = (
                         "type": "object",
                         "properties": {
                             "col1": {
-                                "type": "string"
+                                "type": ["null", "string"]
                             },
                             "_ab_source_file_last_modified": {
                                 "type": "string"
@@ -487,7 +487,7 @@ parquet_file_with_decimal_as_string_scenario = (
     .set_file_type("parquet")
     .set_expected_records(
         [
-            {"data": {"col1": "13.00", "_ab_source_file_last_modified": "2023-06-05T03:54:07Z",
+            {"data": {"col1": "13.00", "_ab_source_file_last_modified": "2023-06-05T03:54:07.000000Z",
                       "_ab_source_file_url": "a.parquet"}, "stream": "stream1"},
         ]
     )
@@ -500,7 +500,7 @@ parquet_file_with_decimal_as_string_scenario = (
                         "type": "object",
                         "properties": {
                             "col1": {
-                                "type": "string"
+                                "type": ["null", "string"]
                             },
                             "_ab_source_file_last_modified": {
                                 "type": "string"
@@ -544,7 +544,7 @@ parquet_file_with_decimal_as_float_scenario = (
     .set_file_type("parquet")
     .set_expected_records(
         [
-            {"data": {"col1": 13.00, "_ab_source_file_last_modified": "2023-06-05T03:54:07Z",
+            {"data": {"col1": 13.00, "_ab_source_file_last_modified": "2023-06-05T03:54:07.000000Z",
                       "_ab_source_file_url": "a.parquet"}, "stream": "stream1"},
         ]
     )
@@ -557,7 +557,7 @@ parquet_file_with_decimal_as_float_scenario = (
                         "type": "object",
                         "properties": {
                             "col1": {
-                                "type": "number"
+                                "type": ["null", "number"]
                             },
                             "_ab_source_file_last_modified": {
                                 "type": "string"
@@ -598,7 +598,7 @@ parquet_file_with_decimal_legacy_config_scenario = (
     .set_file_type("parquet")
     .set_expected_records(
         [
-            {"data": {"col1": 13.00, "_ab_source_file_last_modified": "2023-06-05T03:54:07Z",
+            {"data": {"col1": 13.00, "_ab_source_file_last_modified": "2023-06-05T03:54:07.000000Z",
                       "_ab_source_file_url": "a.parquet"}, "stream": "stream1"},
         ]
     )
@@ -611,7 +611,7 @@ parquet_file_with_decimal_legacy_config_scenario = (
                         "type": "object",
                         "properties": {
                             "col1": {
-                                "type": "number"
+                                "type": ["null", "number"]
                             },
                             "_ab_source_file_last_modified": {
                                 "type": "string"