+++ airbyte-ci/connectors/pipelines/pipelines/pipelines/metadata.py
@@ -63,6 +63,10 @@ class MetadataValidation(PoetryRun):
 
 
 class MetadataUpload(PoetryRun):
+
+    # When the metadata service exits with this code, it means the metadata is valid but the upload was skipped because the metadata is already uploaded
+    skipped_exit_code = 5
+
     def __init__(
         self,
         context: PipelineContext,
@@ -125,7 +129,7 @@ class DeployOrchestrator(Step):
 
     async def _run(self) -> StepResult:
         parent_dir = self.context.get_repo_dir(METADATA_DIR)
-        python_base = with_python_base(self.context)
+        python_base = with_python_base(self.context, "3.9")
         python_with_dependencies = with_pip_packages(python_base, ["dagster-cloud==1.2.6", "pydantic==1.10.6", "poetry2setup==1.1.0"])
         dagster_cloud_api_token_secret: dagger.Secret = get_secret_host_variable(
             self.context.dagger_client, "DAGSTER_CLOUD_METADATA_API_TOKEN"