+++ airbyte-integrations/bases/base-typing-deduping/src/test/java/io/airbyte/integrations/base/destination/typing_deduping/StreamIdTest.java
@@ -4,6 +4,7 @@
 
 package io.airbyte.integrations.base.destination.typing_deduping;
 
+import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
@@ -19,8 +20,16 @@ class StreamIdTest {
     String stream1 = StreamId.concatenateRawTableName("aaa_abab_bbb", "ccc");
     String stream2 = StreamId.concatenateRawTableName("aaa", "bbb_abab_ccc");
 
-    assertEquals("aaa_abab_bbb_ab__ab_ccc", stream1);
-    assertEquals("aaa_ab__ab_bbb_abab_ccc", stream2);
+    assertAll(
+        () -> assertEquals("aaa_abab_bbb_raw__stream_ccc", stream1),
+        () -> assertEquals("aaa_raw__stream_bbb_abab_ccc", stream2));
+  }
+
+  @Test
+  public void noUnderscores() {
+    String stream = StreamId.concatenateRawTableName("a", "b");
+
+    assertEquals("a_raw__stream_b", stream);
   }
 
 }