+++ airbyte-integrations/bases/connector-acceptance-test/acceptance-test-docker.sh
@@ -17,6 +17,7 @@ CONNECTOR_DIR="$ROOT_DIR/airbyte-integrations/connectors/$CONNECTOR_NAME"
 if [ -n "$FETCH_SECRETS" ]; then
   cd $ROOT_DIR
   pip install pipx
+  pipx ensurepath
   pipx install airbyte-ci/connectors/ci_credentials
   VERSION=dev ci_credentials $CONNECTOR_NAME write-to-storage || true
   cd -