+++ airbyte-cdk/python/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.47.5
+Support many format for cursor datetime
+
+## 0.47.4
+File-based CDK updates
+
+## 0.47.3
+Connector Builder: Ensure we return when there are no slices
+
 ## 0.47.2
 low-code: deduplicate query params if they are already encoded in the URL